Фикстуры вынесены в conftest
