+++ tests/test_database.py
@@ -5,13 +5,6 @@ from praktikum.bun import Bun
 from praktikum.ingredient import Ingredient
 from praktikum.ingredient_types import INGREDIENT_TYPE_SAUCE, INGREDIENT_TYPE_FILLING
 
-
-@pytest.fixture
-def database():
-    """Создаем экземпляр базы данных для тестов."""
-    return Database()
-
-
 @allure.feature("База данных")
 @allure.story("Доступные булочки")
 class TestDatabase:
